+++ hata/discord/integration/integration_application.py
@@ -55,8 +55,14 @@ class IntegrationApplication(DiscordEntity):
     
     @property
     def summary(self):
+        """
+        `.summary` is deprecated and will be removed in 2022 Jun. Please use `.description` instead.
+        """
         warnings.warn(
-            f'`{self.__class__.__name__}.summary` is deprecated and will be removed in 2022 Jun.',
+            (
+                f'`{self.__class__.__name__}.summary` is deprecated and will be removed in 2022 Jun. '
+                f'Please use `.description` instead.'
+            ),
             FutureWarning,
         )