+++ hata/ext/prettyprint/prettyprint.py
@@ -895,7 +895,6 @@ def str_guild(guild, index=None, **kwargs):
     result.append(f'- max presences : {guild.max_presences}', 1)
     result.append(f'- max video channel users : {guild.max_video_channel_users}', 1)
     result.append(f'- preferred locale : {guild.preferred_locale}', 1)
-    result.append(f'- region: {guild.region.value}', 1)
     
     description = guild.description
     if (description is not None):